chore: [BB-123] chore** organized frontend directory (#124)
* organized frontend directory

* modified imports
